Move constants to root
