+++ helpers/trades.py
@@ -2,7 +2,7 @@ import click
 from datetime import datetime
 from gql import gql
 
-from .constants import SEPARATOR, COLOR_LABEL, COLOR_LABEL_DELETED, COLOR_SEPARATOR
+from constants import SEPARATOR, COLOR_LABEL, COLOR_LABEL_DELETED, COLOR_SEPARATOR
 from .tokens import TOKEN_FIELDS_BASIC, to_token
 from .utils import get_graphql_client, to_etherscan_link, format_integer, format_amount_in_weis, format_token_long, format_token_short, debug_query, format_date_time
